requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
behave==1.2.6
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
fastapi==0.115.2
graphviz==0.20.3
h11==0.14.0
idna==3.10
iniconfig==2.0.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==24.1
parse==1.20.2
parse_type==0.6.4
pika==1.3.2
pluggy==1.5.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pytest==8.3.3
pytest-bdd==7.3.0
PyYAML==6.0.2
requests==2.32.3
rich==13.9.2
six==1.16.0
sniffio==1.3.1
sse-starlette==2.1.3
starlette==0.40.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
